[+] Fixed bugs

+++ include/components/healthcomponent.h
@@ -6,9 +6,9 @@
 class HealthComponent : public IComponent {
  private:
   friend class AttackSystem;
-  float health;
+  int health;
 
  public:
   explicit HealthComponent(const int _health) : health(_health) {}
-  float getHealth() const;
+  int getHealth() const;
 };
