+++ src/components/healthcomponent.cpp
@@ -1,5 +1,5 @@
 #include "../include/components/healthcomponent.h"
 
-float HealthComponent::getHealth() const {
+int HealthComponent::getHealth() const {
   return health;
 }
